% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.emputation}
\alias{print.emputation}
\title{Print an emputation fit}
\usage{
\method{print}{emputation}(x, ...)
}
\arguments{
\item{x}{An object of class `"emputation"`.}

\item{...}{Unused; present for S3 method consistency.}
}
\value{
`x`, invisibly.
}
\description{
Print an emputation fit
}
